fix: add client and server scripts to custom app and remove related patches
- Added a custom field "custom_sales_invoice" in Stock Entry to determine the linked Sales Invoice.
- Removed previously created patches for adding client and server scripts related to the Trade In feature.
- Updated the custom app to include the required scripts directly.
