+++ csf_tz/custom_api.py
@@ -2755,3 +2755,139 @@ def target_warehouse_based_price_list(doc, method):
             item.price_list_rate = rate
             item.rate = rate
             item.amount = item.qty * rate
+
+def validate_trade_in_serial_no_and_batch(doc, method):
+    error_messages = []
+    for row in doc.items:
+        if row.item_code == "Trade In" and row.custom_trade_in_item:
+            has_batch_no = frappe.db.get_value("Item", row.custom_trade_in_item, "has_batch_no")
+            if has_batch_no and not row.custom_trade_in_batch_no:
+                error_messages.append(f"Batch No. is mandatory for Item {row.custom_trade_in_item} in row {row.idx}.")
+            
+            has_serial_no = frappe.db.get_value("Item", row.custom_trade_in_item, "has_serial_no")
+            if has_serial_no:
+                if not row.custom_trade_in_serial_no:
+                    error_messages.append(f"Serial Numbers are mandatory for Item {row.custom_trade_in_item} in row {row.idx}.")
+                else:
+                    serial_numbers = row.custom_trade_in_serial_no.split("\n")
+                    if len(serial_numbers) != row.custom_trade_in_qty:
+                        error_messages.append(
+                            f"Serial Numbers count ({len(serial_numbers)}) does not match "
+                            f"the Trade-In Quantity ({row.custom_trade_in_qty}) for Item {row.custom_trade_in_item} in row {row.idx}."
+                        )
+    if error_messages:
+        frappe.throw(
+            title="Validation Errors",
+            msg="<br>".join(error_messages),
+        )
+
+
+def validate_trade_in_sales_percentage(doc, method):
+    # Calculate the total trade-in value from the child table where item_code = "Trade In"
+    total_trade_in_value = sum(
+        row.custom_total_trade_in_value for row in doc.items if row.item_code == "Trade In"
+    )
+
+    # Calculate the total for items in the child table where item_code != "Trade In" using the "amount" field
+    non_trade_in_total = sum(
+        row.amount for row in doc.items if row.item_code != "Trade In"
+    )
+
+    # Fetch allowed percentage from the Company doctype
+    trade_in_percentage = frappe.db.get_value("Company", doc.company, "custom_trade_in_sales_percentage") or 0
+
+    # Calculate the allowed trade-in value based on the percentage of non-trade-in total
+    allowed_trade_in_value = (trade_in_percentage / 100) * non_trade_in_total
+
+    # Validate total trade-in value
+    if total_trade_in_value > allowed_trade_in_value:
+        # Throw error if child table total exceeds the allowed limit
+        frappe.throw(
+            title="Trade-In Value Validation Error",
+            msg=(
+                f"Total Trade-In Value ({frappe.format_value(total_trade_in_value)}) exceeds the allowed limit. "
+                f"The allowed limit is {trade_in_percentage}% of the total value of non-Trade-In items in the Sales Invoice. "
+                f"Maximum allowed Trade-In Value: {frappe.format_value(allowed_trade_in_value)}."
+            )
+        )
+
+def create_trade_in_stock_entry(doc, method):
+    # Initialize an empty list to store items
+    items_list = []
+
+    # Fetch Company's Trade_in_control_account and abbreviation
+    company_details = frappe.db.get_value(
+        'Company',
+        doc.company,
+        ['custom_trade_in_control_account', 'abbr'],
+        as_dict=True
+    )
+    if not company_details:
+        frappe.throw(f"Company details not found for {doc.company}. Please check the Company configuration.")
+
+    trade_in_control_account = company_details.get('custom_trade_in_control_account')
+    company_abbr = company_details.get('abbr')
+
+    if not trade_in_control_account:
+        frappe.throw(f"Trade-In Control Account not configured for {doc.company}. Please set it in the Company settings.")
+
+    # Dynamically set warehouse name using company abbreviation
+    to_warehouse = f"Stores - {company_abbr}"
+
+    if not frappe.db.exists("Warehouse", to_warehouse):
+        frappe.throw(f"Warehouse '{to_warehouse}' does not exist. Please create it first.")
+
+    # Iterate through the items in the document
+    for item in doc.items:
+        if item.get("custom_trade_in_item") and item.get("custom_trade_in_qty"):
+            # Check if custom_trade_in_batch_no exists
+            custom_batch_no = item.get("custom_trade_in_batch_no")
+
+            if custom_batch_no:
+                # Check if a Batch with this ID already exists
+                batch_exists = frappe.db.exists("Batch", {"batch_id": custom_batch_no})
+                if not batch_exists:
+                    try:
+                        # Create a new batch with the given custom_trade_in_batch_no
+                        batch_doc = frappe.new_doc("Batch")
+                        batch_doc.item = item.get("custom_trade_in_item")
+                        batch_doc.batch_id = custom_batch_no  # Use the provided custom batch number
+                        batch_doc.save()
+                    except Exception as e:
+                        frappe.throw(f"Error creating Batch: {str(e)}")
+
+            # Append each item's details to the items_list
+            items_list.append({
+                "item_code": item.get("custom_trade_in_item"),
+                "qty": item.get("custom_trade_in_qty"),
+                "uom": item.get("uom") or "Nos",  # Default to "Nos" if UOM is not provided
+                "basic_rate": item.get("custom_trade_in_incoming_rate"),
+                "batch_no": custom_batch_no,  # Use the custom batch number here
+                "serial_no": item.get("custom_trade_in_serial_no"),  # Get custom serial number value
+                "expense_account": trade_in_control_account,
+            })
+
+    # Create a single stock entry if there are items to add
+    if items_list:
+        try:
+            stock_entry = frappe.get_doc({
+                "doctype": "Stock Entry",
+                "stock_entry_type": "Material Receipt",
+                "to_warehouse": to_warehouse,  # Use dynamic warehouse name
+                "items": items_list,  # Use the populated list here
+                "custom_sales_invoice": doc.name,  # Link to the parent Sales Invoice
+            })
+
+            # Insert and submit the Stock Entry
+          
+            stock_entry.insert()
+            stock_entry.submit()
+            frappe.db.commit()
+
+            # Notify the user
+            frappe.msgprint(f"Stock Entry {stock_entry.name} created successfully!")
+        except Exception as e:
+            frappe.db.rollback()
+            frappe.throw(f"Error during Stock Entry creation: {str(e)}")
+    else:
+        frappe.msgprint("No valid items found for stock entry.")
